// src/account/profile/profile.dto.ts
/**
 * @class ProfileDTO
 * @brief DTO that represents a user's profile credentials
 * @member username The user's username
 * @member email The user's email address
 * @member password The user's password
 */
export class ProfileDTO {
    username: string;
    email: string;
    password: string;
}